Move GroupElement to src/components/navigation/
