+++ app/packs/src/components/navigation/GroupElement.js
